dev/containers/python-redis/requirements.txt
black==23.12.1
click==8.1.7
grpcio
grpcio-tools
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
protobuf
tomli==2.0.1
typing_extensions==4.9.0
